% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plotting PCA and Cluster results}
\usage{
plot_pca(res.pca, res.hcpc, show_pca = c("var", "ind", "both"), plot = TRUE)
}
\arguments{
\item{res.pca}{Output of the PCA function.}

\item{res.hcpc}{Output of the HCPC function.}

\item{show_pca}{Choose to plot \code{var}, \code{ind} or \code{both}.}

\item{plot}{A logical value.}
}
\description{
This function plot the results of the PCA and clustering analysis.
}
\examples{
plot_pca(res.pca, res.hcpc, show_pca = "var")
plot_pca(res.pca, res.hcpc, show_pca = "ind")
plot_pca(res.pca, res.hcpc, show_pca = "both")

}
\author{
Johan Ninanya, Javier Rinza
}
