% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thiry.R
\name{Thiry}
\alias{Thiry}
\title{Productivity and resilience capacity indexes}
\usage{
Thiry(id, ys, yp, data)
}
\arguments{
\item{id}{Row name of the data frame data.}

\item{ys}{Column name of the yield data of drought stress condition.}

\item{yp}{Column name of the yield data of potential condition.}

\item{data}{A data frame.}
}
\value{
It returns a list including:
\itemize{\code{indexes}} A data frame of the drought stress indexes SSI, TOL, MP, GMP, and STI.
\itemize{\code{scores}} A data frame of the Thiry's scores.
\itemize{\code{corr1.}} A matrix of Pearson correlation among the drought stress indexes and the Thiry's scores.
\itemize{\code{range}} A list object with the range of the Thiry's scores SSI, TOL, MP, GMP, and STI.
\itemize{\code{comb1.}} A data frame of combinations of the Thiry's scores.
\itemize{\code{corr2.}} A matrix of Pearson correlation among Thiry's combinations and the yield data (ys and yp).
\itemize{\code{comb2.}} A data frame of the best Thiry's combination.
}
\description{
This function calculates the productivity and resilience capacity
following Thiry et al. (2016).
}
\details{
Use \code{inv = TRUE} in the ThiryScore function only for the SSI and TOL indices.
}
\examples{
n <- 20
id <- 1:n
ys <- runif(n)*5
yp <- runif(n)*10
data <- data.frame(id, ys, yp)
ts <- Thiry("id","ys","yp",data)

}
\references{
Thiry, A. A., Chavez Dulanto, P. N., Reynolds, M. P., & Davies, W. J.
(2016). How can we improve crop genotypes to increase stress
resilience and productivity in a future climate? A new crop screening
 method based on productivity and resistance to abiotic stress.
 Journal of experimental botany, 67(19), 5593-5603.
}
\seealso{
\code{ThiryScore}
}
\author{
Johan Ninanya, Javier Rinza
}
