% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThiryScore.R
\name{ThiryScore}
\alias{ThiryScore}
\title{Thiry scores computation}
\usage{
ThiryScore(x, inv = FALSE)
}
\arguments{
\item{x}{A vector of values of some of the yield and tolerance
indeces suggested by Thiry et al. (2016).}

\item{inv}{A logical value. If \code{TRUE} data values will be
inverted, i.e., higher values will turn into lower values,
and vice versa. Defaul to \code{FALSE}.}
}
\value{
It returns a list including:
\itemize{\code{score}} A data frame object with the original value
(\code{x}) and its respective Thiry score (\code{score}).
\itemize{\code{order}} A data frame object.
\itemize{\code{range}} A data frame with the amount of data per score.
\itemize{\code{corr}} Correlation between the index and its score.
}
\description{
This function calculates the Thiry scores (Thiry et al. 2016)
from yield and tolerance indices data (stress susceptibility index
(SSI), tolerance index (TOL), mean productivity index (MP),
geometric mean productivity index (GMP), and stress tolerance
index (STI)).
}
\details{
Use \code{inv = TRUE} only for the SSI and TOL indices.
}
\examples{
x <- runif(50)
ts <- ThiryScore(x)
ts$range
ts <- ThiryScore(x, inv = TRUE)
ts$range

}
\references{
Thiry, A. A., Chavez Dulanto, P. N., Reynolds, M. P., & Davies, W. J.
(2016). How can we improve crop genotypes to increase stress
resilience and productivity in a future climate? A new crop screening
 method based on productivity and resistance to abiotic stress.
 Journal of experimental botany, 67(19), 5593-5603.
}
\seealso{
\code{Thiry}
}
\author{
Johan Ninanya, Javier Rinza
}
